extract github-proxy main to shared package for external invocation (#45113)
This is a copy-and-paste of the github-proxy main function to a shared package so that github-proxy can be invoked from a different entrypoint. This is how it's done for several other programs (frontend, repo-updater, gitserver, etc.). There are no logic changes.
