+++ cmd/github-proxy/shared/shared_test.go
@@ -1,4 +1,4 @@
-package main
+package shared
 
 import (
 	"context"
